requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
groq
python-multipart
pydantic
python-dotenv
cryptography
